pyproject.toml
[tool.poetry]
name = "crowdcomms-django-test"
version = "0.1.0"
description = ""
authors = ["Chilling Chua <chillingchua@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "2.2.13"
djangorestframework = "3.11.0"
pytz = "2018.9"

[tool.poetry.dev-dependencies]
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
